// src/enums/code_category.enum.ts
export enum CodeCategory {
    // 권한
    PERMISSION = "permission",
    // 공통
    COMMON_TYPE = "common_type",
    // 첨부파일
    ATTACHMENT_TYPE = "attachment_type",
    // 특별 역할
    SPECIAL_ROLE_TYPE = "special_role_type",
    // 제품
    PRODUCT_PURPOSE = "product_purpose",
    PRODUCT_CLASS = "product_class",
    PRODUCT_UNIT = "product_unit",
    // 본사 & 거래처
    COMPANY_TYPE = "company_type",
    // 담당자
    EMPLOYEE_ROLE = "employee_role",
    // 계약
    ISSUANCE_METHOD_TYPE = "issuance_method_type",
    BILLING_CYCLE_TYPE = "billing_cycle_type",
    // 유지보수 사업
    INSPECT_CYCLE_TYPE = "inspect_cycle_type",
    INSPECT_METHOD_TYPE = "inspect_method_type",
    // 검수&매출
    PARTNER_TYPE = "partner_type",
}